+++ cpp06/ex01/Serializer.hpp
@@ -6,13 +6,14 @@
 /*   By: cmeng <cmeng@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/09 17:07:03 by cmeng             #+#    #+#             */
-/*   Updated: 2024/01/11 07:11:21 by cmeng            ###   ########.fr       */
+/*   Updated: 2024/01/13 16:19:44 by cmeng            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef SERIALIZER_HPP
 #define SERIALIZER_HPP
 
+#include <stdint.h>
 #include <string>
 
 struct Data {
@@ -22,8 +23,8 @@ struct Data {
 
 class Serializer {
    public:
-    static unsigned long serialize(Data *ptr);
-    static Data *deserialize(unsigned long raw);
+    static uintptr_t serialize(Data *ptr);
+    static Data *deserialize(uintptr_t raw);
 
    private:
     Serializer(void);                                     // Default Constructor
